Information on Primary Stage Units (PSUs) stratification

Example data frame containing information on Primary Stage Units (PSUs) stratification.

data(beat.example)

Format

The PSU_strat data frame contains a row for each Primary Stage Units (PSUs) with the following variables:

STRATUM

Identifier of the stratum (numeric)

PSU_MOS

Measure of size of the primary stage unit (numeric)

PSU_ID

Identifier of the primary stage unit (numeric)

Details

Note: the names of the variables must be the ones indicated above.

Examples

# Load example data
data(beat.example)
PSU_strat

#>     STRATUM PSU_MOS PSU_ID
#> 1         1    2591      1
#> 2         1    3808      2
#> 3         1     465      3
#> 4         1    1778      4
#> 5         1     713      5
#> 6         1    6378      6
#> 7         1     271      7
#> 8         1   17097      8
#> 9         1    2010      9
#> 10        1     521     10
#> 11        1     882     11
#> 12        1    1057     12
#> 13        1   12367     13
#> 14        1    1375     14
#> 15        1     819     15
#> 16        1    3178     16
#> 17        1     551     17
#> 18        1    3825     18
#> 19        1      97     19
#> 20        1    3355     20
#> 21        1    1632     21
#> 22        1    3273     22
#> 23        1     597     23
#> 24        1   18185     24
#> 25        1    3399     25
#> 26        1     566     26
#> 27        1    2088     27
#> 28        1   13502     28
#> 29        1     556     29
#> 30        1    3780     30
#> 31        1     845     31
#> 32        1    2372     32
#> 33        1    3101     33
#> 34        1    8297     34
#> 35        1    4454     35
#> 36        1     472     36
#> 37        1     481     37
#> 38        1    8520     38
#> 39        1    1760     39
#> 40        1    1540     40
#> 41        1    1411     41
#> 42        1    1267     42
#> 43        1    1571     43
#> 44        1    6521     44
#> 45        1    6458     45
#> 46        1    3585     46
#> 47        1    7679     47
#> 48        1    6304     48
#> 49        1    1385     49
#> 50        1    1317     50
#> 51        1    5591     51
#> 52        1     301     52
#> 53        1    2588     53
#> 54        1     554     54
#> 55        1    2136     55
#> 56        1    1008     56
#> 57        1     771     57
#> 58        1    9234     58
#> 59        1   28653     59
#> 60        1    1851     60
#> 61        1    1506     61
#> 62        1    2874     62
#> 63        1   18577     63
#> 64        1    1820     64
#> 65        1    2186     65
#> 66        1   10008     66
#> 67        1     415     67
#> 68        1    6336     68
#> 69        1    2383     69
#> 70        1    5642     70
#> 71        1    1864     71
#> 72        1    1080     72
#> 73        1     164     73
#> 74        1    1040     74
#> 75        1     360     75
#> 76        1    1695     76
#> 77        1    2150     77
#> 78        1   36168     78
#> 79        1     216     79
#> 80        1     942     80
#> 81        1    1687     81
#> 82        1   26368     82
#> 83        1     365     83
#> 84        1     264     84
#> 85        1     389     85
#> 86        1   18973     86
#> 87        1     212     87
#> 88        1    1571     88
#> 89        1    3339     89
#> 90        1   50137     90
#> 91        1     358     91
#> 92        1     610     92
#> 93        1    4704     93
#> 94        1    3410     94
#> 95        1     528     95
#> 96        1    1000     96
#> 97        1    7882     97
#> 98        1   10141     98
#> 99        1    8513     99
#> 100       1     270    100
#> 101       1    5320    101
#> 102       1    2410    102
#> 103       1     571    103
#> 104       1    2744    104
#> 105       1     859    105
#> 106       1    2372    106
#> 107       1    3691    107
#> 108       1     277    108
#> 109       1    1753    109
#> 110       1    2870    110
#> 111       1     556    111
#> 112       1    9553    112
#> 113       1    1293    113
#> 114       1     661    114
#> 115       1   16730    115
#> 116       1    3509    116
#> 117       1     745    117
#> 118       1     205    118
#> 119       1    1055    119
#> 120       1   37870    120
#> 121       1      47    121
#> 122       1     743    122
#> 123       1     394    123
#> 124       1     422    124
#> 125       1   24196    125
#> 126       1    1821    126
#> 127       1    8457    127
#> 128       1    5303    128
#> 129       1    1546    129
#> 130       1   15451    130
#> 131       1     190    131
#> 132       1    1994    132
#> 133       1     459    133
#> 134       1    1642    134
#> 135       1    1712    135
#> 136       1    1075    136
#> 137       1    1106    137
#> 138       1     380    138
#> 139       1    7676    139
#> 140       1     531    140
#> 141       1     561    141
#> 142       1    1247    142
#> 143       1     467    143
#> 144       1    1409    144
#> 145       1      61    145
#> 146       1    4074    146
#> 147       1     715    147
#> 148       1    4177    148
#> 149       1     903    149
#> 150       1    1279    150
#> 151       1      96    151
#> 152       1     844    152
#> 153       1     411    153
#> 154       1     669    154
#> 155       1     396    155
#> 156       1   58320    156
#> 157       1      42    157
#> 158       1     714    158
#> 159       1     990    159
#> 160       1    3493    160
#> 161       1    5405    161
#> 162       1     897    162
#> 163       1     810    163
#> 164       1   48946    164
#> 165       1     185    165
#> 166       1    6907    166
#> 167       1     320    167
#> 168       1    8034    168
#> 169       1     576    169
#> 170       1    1448    170
#> 171       1   22345    171
#> 172       1     820    172
#> 173       1    1138    173
#> 174       1     897    174
#> 175       1    3209    175
#> 176       1    1258    176
#> 177       1     850    177
#> 178       1    2019    178
#> 179       1     479    179
#> 180       1    1107    180
#> 181       1    3899    181
#> 182       1     218    182
#> 183       1    3954    183
#> 184       1    3438    184
#> 185       1     578    185
#> 186       1     737    186
#> 187       1     772    187
#> 188       1     614    188
#> 189       1   14088    189
#> 190       1    3073    190
#> 191       1   36158    191
#> 192       1    8648    192
#> 193       1    3723    193
#> 194       1   18279    194
#> 195       1    3435    195
#> 196       1    1381    196
#> 197       1   10245    197
#> 198       1    1087    198
#> 199       1    3756    199
#> 200       1    1124    200
#> 201       1     794    201
#> 202       1     266    202
#> 203       1    1939    203
#> 204       1     248    204
#> 205       1    1302    205
#> 206       1     798    206
#> 207       1     564    207
#> 208       1     339    208
#> 209       1     370    209
#> 210       1    1065    210
#> 211       1    1688    211
#> 212       1      69    212
#> 213       1    1162    213
#> 214       1   19422    214
#> 215       1    4199    215
#> 216       1    2721    216
#> 217       1   12370    217
#> 218       1    1658    218
#> 219       1   49591    219
#> 220       1    3023    220
#> 221       1    1781    221
#> 222       1    2028    222
#> 223       1    2957    223
#> 224       1     319    224
#> 225       1    1843    225
#> 226       1     251    226
#> 227       1     897    227
#> 228       1    4623    228
#> 229       1    2399    229
#> 230       1     765    230
#> 231       1    1771    231
#> 232       1     572    232
#> 233       1     543    233
#> 234       1      79    234
#> 235       1    1661    235
#> 236       1    5675    236
#> 237       1    3943    237
#> 238       1     375    238
#> 239       1     574    239
#> 240       1    4877    240
#> 241       1    3777    241
#> 242       1    1878    242
#> 243       1    3035    243
#> 244       1    2721    244
#> 245       1    1041    245
#> 246       1    3383    246
#> 247       1     865    247
#> 248       1    9763    248
#> 249       1   19311    249
#> 250       1    1471    250
#> 251       1     285    251
#> 252       1    3122    252
#> 253       1    1920    253
#> 254       1    3615    254
#> 255       1    4843    255
#> 256       1    4407    256
#> 257       1   10740    257
#> 258       1     252    258
#> 259       1    1171    259
#> 260       1    3311    260
#> 261       1     811    261
#> 262       1    1515    262
#> 263       1     889    263
#> 264       1     523    264
#> 265       1   47790    265
#> 266       1    1576    266
#> 267       1    1115    267
#> 268       1     273    268
#> 269       1    6437    269
#> 270       1    6727    270
#> 271       1     807    271
#> 272       1  907563    272
#> 273       1    2748    273
#> 274       1     591    274
#> 275       1    4735    275
#> 276       1    3874    276
#> 277       1     135    277
#> 278       1     350    278
#> 279       1     564    279
#> 280       1   11018    280
#> 281       1     191    281
#> 282       1     224    282
#> 283       1    1472    283
#> 284       1    3850    284
#> 285       1     952    285
#> 286       1     804    286
#> 287       1    3176    287
#> 288       1     113    288
#> 289       1     829    289
#> 290       1    1496    290
#> 291       1     959    291
#> 292       1   34859    292
#> 293       1    5037    293
#> 294       1    1472    294
#> 295       1     834    295
#> 296       1     247    296
#> 297       1     885    297
#> 298       1     499    298
#> 299       1    5268    299
#> 300       1    4845    300
#> 301       1    1141    301
#> 302       1    3334    302
#> 303       1    3043    303
#> 304       1    1015    304
#> 305       1    2079    305
#> 306       1    1124    306
#> 307       1    4184    307
#> 308       1    4898    308
#> 309       1   14200    309
#> 310       1    1212    310
#> 311       1    1342    311
#> 312       1     547    312
#> 313       1    1141    313
#> 314       1   15097    314
#> 315       1    8707    315
#> 316       2     434      2
#> 317       2     342      3
#> 318       2    2716      4
#> 319       2     939      6
#> 320       2    1390      7
#> 321       2      92      8
#> 322       2     250      9
#> 323       2    2060     11
#> 324       2     221     14
#> 325       2    2644     15
#> 326       2   13336     16
#> 327       2    2372     17
#> 328       2     188     19
#> 329       2     922     21
#> 330       2     246     25
#> 331       2      79     29
#> 332       2    1051     30
#> 333       2    1145     31
#>  [ reached 'max' / getOption("max.print") -- omitted 1988 rows ]

str(PSU_strat)

#> 'data.frame':	2321 obs. of  3 variables:
#>  $ STRATUM: int  1 1 1 1 1 1 1 1 1 1 ...
#>  $ PSU_MOS: int  2591 3808 465 1778 713 6378 271 17097 2010 521 ...
#>  $ PSU_ID : int  1 2 3 4 5 6 7 8 9 10 ...