Compute multivariate optimal allocation for different domains in one-stage stratified sample design

This function computes the multivariate optimal allocation for different domains in a one-stage file_strataied sample design, following the generalized Bethel algorithm.

beat.1st(file_strata, errors, var_label=NULL, minnumstrat = 2, maxiter = 200, maxiter1 = 25, epsilon = 1e-11)

Arguments

file_strata:

dataframe of survey strata, for more details see, e.g.,strata.

errors:

dataframe of expected coefficients of variation (CV) for each domain, for more details see, e.g.,errors.

var_label:

a vector listing the target variable(s) which are used for leading the optimal allocation.

minnumstrat:

Minimum number of elementary units per strata (default=2).

maxiter:

Maximum number of iterations (default=200) of the general procedure. This kind of iteration may be required by the fact that when in a stratum the number of allocated units is greater or equal to its population, that stratum is set as "census stratum", and the whole procedure is re-initialised.

maxiter1:

Maximum number of iterations in Chromy algorithm (default=25).

epsilon:

Tollerance for the maximum absolute differences between the expected CV and the realised CV with the allocation obtained in the last iteraction for all domains. The default is 10^(-11).

Value

It returns a list with the following objects:

- a dataframe, *file_strata*, with the strata information that is one of the inputs for performing the multivariate and multidomain optimal allocation implemented in the function *beat.1st* in R2BEAT R-package;

- a vector, *var_label*, listing the target variable(s) which will be used for leading the optimal allocation.

Returns a list with four components: - n: Vector with the optimal sample size for each stratum

- file_strata: dataframe corresponding to the input data.frame file_strata with the optimal sample size column added

- alloc: dataframe with optimal (OPT), proportional (PROP), and equal (UNIF) saple size allocation.

- sensitivity: dataframe with the planned coefficients of variation (PlannedCV), the expected coefficients of variation (ExpectedCV), and the sensitivity at 10

- expectedCV: dataframe with the maximum of the expected coefficients of variation (Expected CV), for each variable in each domain.

- plannedCV: dataframe with the maximum coefficients of variation admissible for each domain and for each variable. It is the input, error, provided by the user.

- param_alloc: a vector summarising all the parameters used for performing the optimal allocation.

Details

The methodology is a generalization of Bethel multivariate allocation (1989) that extends the Neyman (1959) - Tchuprov (1923) allocation for multi-purpose and multi-domain surveys. The generalized Bethel’s algorithm allows to determine the optimal sample size for each stratum in a stratified sample design. The overall sample size and the allocation among the different strata is determined starting from the accuracy constraints imposed in the survey on interest estimates.

Examples

# Load example data
data(beat.example)
allocation_1 <- beat.1st(file_strata=strata, errors=errors)
sum(allocation_1$n)
#> [1] 6858
allocation_1$sensitivity
#>    Type Dom Dom_label Var PlannedCV ExpectedCV Sensitivity10%
#> 1  DOM1   1      REG1  V1      0.10     0.1000            925
#> 2  DOM1   1      REG1  V2      0.99     0.1100              1
#> 3  DOM1   2      REG2  V1      0.10     0.0999            446
#> 4  DOM1   2      REG2  V2      0.99     0.1718              1
#> 5  DOM2   1     PROV1  V1      0.99     0.1044              1
#> 6  DOM2   1     PROV1  V2      0.99     0.1151              1
#> 7  DOM2   2     PROV2  V1      0.99     0.3477              1
#> 8  DOM2   2     PROV2  V2      0.99     0.3740              1
#> 9  DOM2   3     PROV3  V1      0.99     0.3253              1
#> 10 DOM2   3     PROV3  V2      0.99     0.4718              1
#> 11 DOM2   4     PROV4  V1      0.99     0.4599              1
#> 12 DOM2   4     PROV4  V2      0.99     0.8582              1
#> 13 DOM2   5     PROV5  V1      0.99     0.2024              1
#> 14 DOM2   5     PROV5  V2      0.99     0.3717              1
#> 15 DOM2   6     PROV6  V1      0.99     0.1268              1
#> 16 DOM2   6     PROV6  V2      0.99     0.2186              1
errors1 <- errors 
errors1[1,2] <- errors[1,2]+errors[1,2]*0.1
allocation_2 <- beat.1st(file_strata=strata, errors=errors1)
sum(allocation_2$n)
#> [1] 5671
errors2 <- errors 
errors2[1,2] <- errors[1,2]-errors[1,2]*0.1
allocation_3 <- beat.1st(file_strata=strata, errors=errors2)
sum(allocation_3$n)
#> [1] 8463