Compute multivariate optimal allocation for different domains in two-stage stratified sample design

This function computes the multivariate optimal allocation for different domains in a one-stage file_strataied sample design, following the generalized Bethel algorithm.

beat.1st(file_strata,errors,des_file,psu_file,rho,deft_start,effst,epsilon1,mmdiff_deft,maxi,epsilon,minPSUstrat,minnumstrat,maxiter,maxiter1)

Arguments

file_strata:

dataframe of survey strata, for more details see, e.g.,strata.

errors:

dataframe of coefficients of variation for each domain, for more details see, e.g., errors.

des_file:

dataframe containing information on sampling design variables, for more details see, e.g., design.

psu_file:

dataframe containing information on primary stage units stratification, for more details see, e.g., PSU_strat.

rho:

dataframe of survey strata, for more details see, e.g., rho.

deft_start:

dataframe of survey strata, for taking into account the initial design effect on each variable, for more details see, e.g., deft_start.

effst:

dataframe of survey strata, for taking into account the estimator effect on each variable, for more details see, e.g., effst.

epsilon1:

first stop condition: sample sizes differences beetween two iterations; iteration continues until the maximum of sample sizes differences is greater than the default value. The default is 5.

mmdiff_deft:

second stop condition: defts differences beetween two iterations; iteration continues until the maximum of defts largest differences is greater than the default value. The default is 0.06.

maxi:

third stop condition: maximum number of allowed iterations. The default is 20.

epsilon:

the same as in function beat.1st.

minPSUstrat:

minimum number of non-self-represenative PSUs to be selected in each stratum.

minnumstrat:

the same as in function beat.1st.

maxiter:

the same as in function beat.1st.

maxiter1:

the same as in function beat.1st.

Value

Returns a list with four components:

- iteractions: dataframe that for each iteraction provides a summary with the number of Primary Stage Units (PSU_Total) distinguish between Self-Representative (PSU_SR) from Non-Self-Representative (PSU_NSR) and the number of Secondary Stage Units (SSU). This output is also printed to the screen.

- file_strata: Input dataframe in file_strata with the design effect for each variables in each stratum (DEFT1 - DEFTn) and the optimal sample size columns.

- alloc: dataframe with optimal (OPT), proportional (PROP), equal (UNIF) sample size allocation.

- planned: Data frame with a summary of expected coefficients of variation for each variable in each domain.

- expected: Data frame with a summary of realized coefficients of variation with the given optimal allocation for each variable in each domain.

- sensitivity: Data frame with a summary of the sensitivity at 10% for each domain and each variable. Sensitivity can be a useful tool to help in finding the best allocation, because it provides a hint of the expected sample size variation for a 10% change in planned CVs.

- deft_c: Data frame with the design effect for each variable in each domain in each iteraction. Note that DEFT1_0 - DEFTn_0 is always equal to 1 if deft_start is NULL. Instead is equal to deft_start. While DEFT1 - DEFTn are the final design effect related to the given allocation.

- param_alloc: A vector with a resume of all the parameter given for the allocation.

Examples

# Load example data
data(beat.example)

allocation2st_1 <- beat.2st(file_strata=strata, errors=errors,
                           des_file=design, psu_file=PSU_strat,rho=rho)
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#>   iterations PSU_SR PSU NSR PSU Total   SSU
#> 1          0      0       0         0  6858
#> 2          1     24      88       112 16113
#> 3          2     59     164       223 13771
#> 4          3     46     156       202 14490
#> 5          4     49     158       207 14299

allocation2st_1$sensitivity
#>    Type Dom   V1 V2
#> 1  DOM1   1 1798  1
#> 3  DOM1   2 1061  1
#> 5  DOM2   1    1  1
#> 7  DOM2   2    1  1
#> 9  DOM2   3    1  1
#> 11 DOM2   4    1  1
#> 13 DOM2   5    1  1
#> 15 DOM2   6    1  1
errors1 <- errors 
errors1[1,2] <- errors[1,2]+errors[1,2]*0.1
allocation2st_2 <- beat.2st(file_strata=strata, errors=errors1,
                           des_file=design, psu_file=PSU_strat,rho=rho)
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#>    iterations PSU_SR PSU NSR PSU Total   SSU
#> 1           0      0       0         0  5671
#> 2           1     17      86       103 14044
#> 3           2     51     146       197 11731
#> 4           3     40     138       178 12293
#> 5           4     43     138       181 12153
#> 6           5     40     140       180 12292
#> 7           6     43     138       181 12152
#> 8           7     40     140       180 12293
#> 9           8     43     138       181 12153
#> 10          9     40     140       180 12292
#> 11         10     43     138       181 12152
#> 12         11     40     140       180 12293
#> 13         12     43     138       181 12153
#> 14         13     40     140       180 12292
#> 15         14     43     138       181 12152
#> 16         15     40     140       180 12293
#> 17         16     43     138       181 12153
#> 18         17     40     140       180 12292
#> 19         18     43     138       181 12152
#> 20         19     40     140       180 12293
#> 21         20     43     138       181 12153

errors2 <- errors 
errors2[1,2] <- errors[1,2]-errors[1,2]*0.1
allocation2st_3 <- beat.2st(file_strata=strata, errors=errors2,
                          des_file=design, psu_file=PSU_strat,rho=rho)
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#>    iterations PSU_SR PSU NSR PSU Total   SSU
#> 1           0      0       0         0  8463
#> 2           1     28     106       134 19335
#> 3           2     70     186       256 16525
#> 4           3     60     178       238 17048
#> 5           4     62     180       242 16901
#> 6           5     62     180       242 16895
#> 7           6     64     178       242 16828
#> 8           7     61     180       241 16937
#> 9           8     63     180       243 16862
#> 10          9     63     178       241 16859
#> 11         10     62     180       242 16903
#> 12         11     62     180       242 16895
#> 13         12     64     178       242 16828
#> 14         13     61     180       241 16937
#> 15         14     63     180       243 16862
#> 16         15     63     178       241 16859
#> 17         16     62     180       242 16903
#> 18         17     62     180       242 16895
#> 19         18     64     178       242 16828
#> 20         19     61     180       241 16937
#> 21         20     63     180       243 16862
design1 <- design 
design1[,4] <- 60
allocation2st_4 <- beat.2st(file_strata=strata, errors=errors2,
                           des_file=design1, psu_file=PSU_strat, rho=rho)
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#>    iterations PSU_SR PSU NSR PSU Total   SSU
#> 1           0      0       0         0  8463
#> 2           1     24      88       112 22928
#> 3           2     67     182       249 18703
#> 4           3     52     168       220 19834
#> 5           4     60     168       228 19212
#> 6           5     54     170       224 19654
#> 7           6     58     170       228 19384
#> 8           7     56     168       224 19479
#> 9           8     58     168       226 19385
#> 10          9     56     168       224 19477
#> 11         10     58     168       226 19386
#> 12         11     56     168       224 19478
#> 13         12     58     168       226 19384
#> 14         13     56     168       224 19479
#> 15         14     58     168       226 19385
#> 16         15     56     168       224 19477
#> 17         16     58     168       226 19386
#> 18         17     56     168       224 19478
#> 19         18     58     168       226 19384
#> 20         19     56     168       224 19479
#> 21         20     58     168       226 19385
allocation2st_3$expected
#>    Type Dom     V1     V2
#> 1  DOM1   1 0.0900 0.0740
#> 3  DOM1   2 0.0899 0.1107
#> 5  DOM2   1 0.0931 0.0773
#> 7  DOM2   2 0.3451 0.2593
#> 9  DOM2   3 0.3013 0.3251
#> 11 DOM2   4 0.4658 0.4905
#> 13 DOM2   5 0.1773 0.2467
#> 15 DOM2   6 0.1134 0.1373
allocation2st_4$expected
#>    Type Dom     V1     V2
#> 1  DOM1   1 0.0900 0.0692
#> 3  DOM1   2 0.0900 0.1054
#> 5  DOM2   1 0.0930 0.0721
#> 7  DOM2   2 0.3510 0.2477
#> 9  DOM2   3 0.3052 0.3116
#> 11 DOM2   4 0.4600 0.4693
#> 13 DOM2   5 0.1768 0.2367
#> 15 DOM2   6 0.1133 0.1299
design2 <- design 
design2[,4] <- 24
allocation2st_5 <- beat.2st(file_strata=strata, errors=errors2,
                          des_file=design2, psu_file=PSU_strat, rho=rho)
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#>   iterations PSU_SR PSU NSR PSU Total   SSU
#> 1          0      0       0         0  8463
#> 2          1     62     166       228 12615
#> 3          2     95     220       315 11986
#> 4          3     89     216       305 12096
#> 5          4     90     214       304 12085
allocation2st_4$expected
#>    Type Dom     V1     V2
#> 1  DOM1   1 0.0900 0.0692
#> 3  DOM1   2 0.0900 0.1054
#> 5  DOM2   1 0.0930 0.0721
#> 7  DOM2   2 0.3510 0.2477
#> 9  DOM2   3 0.3052 0.3116
#> 11 DOM2   4 0.4600 0.4693
#> 13 DOM2   5 0.1768 0.2367
#> 15 DOM2   6 0.1133 0.1299
allocation2st_5$expected
#>    Type Dom     V1     V2
#> 1  DOM1   1 0.0900 0.0861
#> 3  DOM1   2 0.0899 0.1270
#> 5  DOM2   1 0.0934 0.0902
#> 7  DOM2   2 0.3342 0.2911
#> 9  DOM2   3 0.2984 0.3593
#> 11 DOM2   4 0.4489 0.5824
#> 13 DOM2   5 0.1783 0.2782
#> 15 DOM2   6 0.1138 0.1602
design
#>    STRATUM STRAT_MOS DELTA MINIMUM
#> 1        1   2302353     1      48
#> 2        2    179562     1      48
#> 3        3    371802     1      48
#> 4        4    592303     1      48
#> 5        5    221687     1      48
#> 6        6    440613     1      48
#> 7        7    128230     1      48
#> 8        8    222648     1      48
#> 9        9    287906     1      48
#> 10      10    882718     1      48
#> 11      11    223516     1      48
#> 12      12    883285     1      48
#> 13      13    594988     1      48
#> 14      14    183169     1      48
#> 15      15   3156694     1      48
#> 16      16   1098740     1      48
#> 17      17   1256025     1      48
design1
#>    STRATUM STRAT_MOS DELTA MINIMUM
#> 1        1   2302353     1      60
#> 2        2    179562     1      60
#> 3        3    371802     1      60
#> 4        4    592303     1      60
#> 5        5    221687     1      60
#> 6        6    440613     1      60
#> 7        7    128230     1      60
#> 8        8    222648     1      60
#> 9        9    287906     1      60
#> 10      10    882718     1      60
#> 11      11    223516     1      60
#> 12      12    883285     1      60
#> 13      13    594988     1      60
#> 14      14    183169     1      60
#> 15      15   3156694     1      60
#> 16      16   1098740     1      60
#> 17      17   1256025     1      60
design2
#>    STRATUM STRAT_MOS DELTA MINIMUM
#> 1        1   2302353     1      24
#> 2        2    179562     1      24
#> 3        3    371802     1      24
#> 4        4    592303     1      24
#> 5        5    221687     1      24
#> 6        6    440613     1      24
#> 7        7    128230     1      24
#> 8        8    222648     1      24
#> 9        9    287906     1      24
#> 10      10    882718     1      24
#> 11      11    223516     1      24
#> 12      12    883285     1      24
#> 13      13    594988     1      24
#> 14      14    183169     1      24
#> 15      15   3156694     1      24
#> 16      16   1098740     1      24
#> 17      17   1256025     1      24
design3 <- design 
design3$DELTA <- 2.31
allocation2st_6 <- beat.2st(file_strata=strata, errors=errors,
                           des_file=design3, psu_file=PSU_strat, rho=rho)
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#> Warning: The number of elements in var_label does not match with the number of variables used in the allocation
#>   iterations PSU_SR PSU NSR PSU Total   SSU
#> 1          0      0       0         0  6858
#> 2          1      9      60        69 37754
#> 3          2     60     152       212 28110
#> 4          3     40     144       184 30383
#> 5          4     48     144       192 29401
#> 6          5     46     140       186 29651
library(readr)
#> Error in library(readr): there is no package called ‘readr’
pop <- read_rds("https://github.com/barcaroli/R2BEAT_workflows/blob/master/pop.RDS?raw=true")
#> Error in read_rds("https://github.com/barcaroli/R2BEAT_workflows/blob/master/pop.RDS?raw=true"): could not find function "read_rds"
library(R2BEAT)
cv <- as.data.frame(list(DOM=c("DOM1","DOM2"),
                        CV1=c(0.02,0.03),
                        CV2=c(0.03,0.06),
                        CV3=c(0.03,0.06),
                        CV4=c(0.05,0.08)))
cv
#>    DOM  CV1  CV2  CV3  CV4
#> 1 DOM1 0.02 0.03 0.03 0.05
#> 2 DOM2 0.03 0.06 0.06 0.08
samp_frame <- pop
#> Error: object 'pop' not found
samp_frame$one <- 1
#> Error: object 'samp_frame' not found
id_PSU <- "municipality"  
id_SSU <- "id_ind"        
strata_var <- "stratum"   
target_vars <- c("income_hh","active","inactive","unemployed")   
deff_var <- "stratum"     
domain_var <- "region"  
delta =  1       # households = survey units
minimum <- 50    # minimum number of SSUs to be interviewed in each selected PSU
deff_sugg <- 1.5 # suggestion for the deff value

inp <- prepareInputToAllocation1(samp_frame,
                               id_PSU,
                                id_SSU,
                                strata_var,
                                target_vars,
                                deff_var,
                                domain_var,
                                minimum,
                                delta,
                                deff_sugg)
#> Error: object 'samp_frame' not found
inp$desfile$MINIMUM <- 50
#> Error: object 'inp' not found
alloc <- beat.2st(file_strata = inp$strata, 
                 errors = cv, 
                 des_file = inp$des_file, 
                 psu_file = inp$psu_file, 
                 rho = inp$rho, 
                 deft_start = NULL,
                 effst = inp$effst, 
                 minPSUstrat = 2,
                 minnumstrat = 50)
#> Error: object 'inp' not found