Multivariate optimal allocation for different domains in two stage statified sample design

Compute multivariate optimal allocation for different domains corrected considering stratified two stages design

beat.2st(stratif, errors, des_file, psu_file, rho, deft_start = NULL,
    effst = NULL, epsilon1 = 5, mmdiff_deft = 1,maxi = 20,
    epsilon = 10^(-11), minnumstrat = 2, maxiter = 200, maxiter1 = 25)

Arguments

stratif
Data frame of survey strata, for more details see, e.g.,strata.

errors
Data frame of coefficients of variation for each domain, for more details see, e.g.,errors.

des_file
Data frame containing information on sampling design variables, for more details see, e.g.,design.

psu_file
Data frame containing information on primary stage units stratification, for more details see, e.g.,PSU_strat.

rho
Data frame of survey strata, for more details see, e.g.,rho.

deft_start
Data frame of survey strata, for taking into account the initial design effect on each variable, for more details see, e.g.,deft_start.

effst
Data frame of survey strata, for taking into account the estimator effect on each variable, for more details see, e.g.,effst.

epsilon1
First stop condition: sample sizes differences beetween two iterations; iteration continues until the maximum of sample sizes differences is greater than the default value. The default is 5.

mmdiff_deft
Second stop condition: defts differences beetween two iterations; iteration continues until the maximum of defts largest differences is greater than the default value. The default is 0.06.

maxi
Third stop condition: maximum number of allowed iterations. The default is 20.

epsilon
The same as in function beat.1st.

minnumstrat
The same as in function beat.1st.

maxiter
The same as in function beat.1st.

maxiter1
The same as in function beat.1st.

Details

The methodology is a generalization of Bethel multivariate allocation (1989) that extended the Neyman (1959) - Tchuprov (1923) allocation for multi-purpose and multi-domains surveys. The generalized Bethel’s algorithm allows to determine the optimal sample size for each stratum in a stratified sample design. The overall sample size and the allocation among the different strata is determined starting from the accuracy constraints imposed in the survey on interest estimates. The optimal allocation is obtained throught a procedure that converge in few iteractions:

The first iteration is a computation of an initial allocation with the multivariate optimal allocation for different domains in one stages statified sample design (the methodology is a generalization for multidomains and multistages designs of Bethel multivariate allocation, 1989).

The correction of the initial allocation is based on an iterative method calculating new allocations and is based on an inflaction of strata variances using the design effect (Ganninger, 2010).

Value

Object of class list. The list contains 8 objects:

iteractions

Data frame that for each iteraction provides a summary with the number of Primary Stage Units (PSU_Total) distinguish between Self-Representative (PSU_SR) from Non-Self-Representative (PSU_NSR) and the number of Secondary Stage Units (SSU). This output is also printed to the screen.

file_strata

Input data frame in stratif with the design effect for each variables in each stratum (DEFT1 - DEFTn) and the optimal sample size columns.

alloc

Data frame with optimal (ALLOC), proportional (PROP), equal (EQUAL) sample size allocation.

planned

Data frame with a summary of expected coefficients of variation for each variable in each domain.

expected

Data frame with a summary of realized coefficients of variation with the given optimal allocation for each variable in each domain.

sensitivity

Data frame with a summary of the sensitivity at 10% for each domain and each variable. Sensitivity can be a useful tool to help in finding the best allocation, because it provides a hint of the expected sample size variation for a 10% change in planned CVs.

deft_c

Data frame with the design effect for each variable in each domain in each iteraction. Note that DEFT1_0 - DEFTn_0 is always equal to 1 if deft_start is NULL. Instead is equal to deft_start. While DEFT1 - DEFTn are the final design effect related to the given allocation.

param_alloc

A vector with a resume of all the parameter given for the allocation.

References

Cochran, W. (1977) Sampling Techniques. John Wiley & Sons, Inc., New York

Ganninger, M. (2010). Design effects: model-based versus design-based approach. Vol. 3, p. 174. DEU.

Neyman, J. (1934). On the two different aspects of the representative method: the method of stratified sampling and the method of purposive selection. Journal of the Royal Statistical Society, 97(4), 558-625.

Tschuprow, A. A. (1923). On the mathematical expectation of the moments of frequency distributions in the case of correlated observation. (Chapters 4-6). Metron, 1923, 2: 646-683.

Examples

# Load example data
data(beat.example)

## Example 1
# Allocate the sample
allocation2st_1 <- beat.2st(stratif=strata, errors=errors,
des_file=design, psu_file=PSU_strat,rho=rho)

#>    iterations PSU_SR PSU NSR PSU Total   SSU
#> 1           0      0       0         0  6858
#> 2           1     16      79        95 17170
#> 3           2     42     161       203 14736
#> 4           3     33     153       186 15273
#> 5           4     36     155       191 15147
#> 6           5     33     156       189 15274
#> 7           6     36     155       191 15146
#> 8           7     33     156       189 15273
#> 9           8     36     155       191 15147
#> 10          9     33     156       189 15274
#> 11         10     36     155       191 15146
#> 12         11     33     156       189 15273
#> 13         12     36     155       191 15147
#> 14         13     33     156       189 15274
#> 15         14     36     155       191 15146
#> 16         15     33     156       189 15273
#> 17         16     36     155       191 15147
#> 18         17     33     156       189 15274
#> 19         18     36     155       191 15146
#> 20         19     33     156       189 15273
#> 21         20     36     155       191 15147

# The total ammount of sample size is 191 PSU (36 SR + 155 NSR) and 15147 SSU. 

## Example 2
# Assume 13000 SSUs is the maximum sample size to stick to our budget.
# Look at the sensitivity is in DOM1 for REG1 and REG2 due to V1.
allocation2st_1$sensitivity

#>    Type Dom   V1 V2
#> 1  DOM1   1 1968  1
#> 3  DOM1   2 1061  1
#> 5  DOM2   1    1  1
#> 7  DOM2   2    1  1
#> 9  DOM2   3    1  1
#> 11 DOM2   4    1  1
#> 13 DOM2   5    1  1
#> 15 DOM2   6    1  1

# We can relax the constraints increasing the expected coefficients of variation for X1 by 10%
errors1 <- errors
errors1[1,2] <- errors[1,2]+errors[1,2]*0.1

# Try the new allocation 
allocation2st_2 <- beat.2st(stratif=strata, errors=errors1,
des_file=design, psu_file=PSU_strat,rho=rho)

#>   iterations PSU_SR PSU NSR PSU Total   SSU
#> 1          0      0       0         0  5671
#> 2          1     13      68        81 14602
#> 3          2     35     143       178 12492
#> 4          3     29     134       163 12944


## Example 3
# On the contrary, if we tighten the constraints decreasing the expected coefficients of variation 
# for X1 by 10%
errors2 <- errors
errors2[1,2] <- errors[1,2]-errors[1,2]*0.1

# The new allocation leads to a larger sample than the first example (around 18000)
allocation2st_3 <- beat.2st(stratif=strata, errors=errors2,
des_file=design, psu_file=PSU_strat,rho=rho)

#>    iterations PSU_SR PSU NSR PSU Total   SSU
#> 1           0      0       0         0  8463
#> 2           1     18      95       113 20796
#> 3           2     55     182       237 17379
#> 4           3     39     175       214 18451
#> 5           4     47     175       222 17957
#> 6           5     41     177       218 18272
#> 7           6     47     174       221 17957
#> 8           7     41     177       218 18272
#> 9           8     47     174       221 17957
#> 10          9     41     177       218 18272
#> 11         10     47     174       221 17957
#> 12         11     41     177       218 18272
#> 13         12     47     174       221 17957
#> 14         13     41     177       218 18272
#> 15         14     47     174       221 17957
#> 16         15     41     177       218 18272
#> 17         16     47     174       221 17957
#> 18         17     41     177       218 18272
#> 19         18     47     174       221 17957
#> 20         19     41     177       218 18272
#> 21         20     47     174       221 17957


## Example 4
# Sometimes some budget constraints concern the number of PSU involved in the survey.
# Tuning the PSUs number is possible modyfing the MINIMUM in des_file. 
# Assume to increase the MINIMUM from 48 to 60
design1 <- design
design1[,4] <- 60
allocation2st_4 <- beat.2st(stratif=strata, errors=errors2,
des_file=design1, psu_file=PSU_strat, rho=rho)

#>   iterations PSU_SR PSU NSR PSU Total   SSU
#> 1          0      0       0         0  8463
#> 2          1     15      79        94 24569
#> 3          2     52     175       227 19798
#> 4          3     38     162       200 21051
#> 5          4     41     166       207 20821
#> 6          5     39     167       206 20971
#> 7          6     40     167       207 20895


# The PSUs numer is decreased, while the SSUs number increased 
# due to cluster intra-correlation effect.
# Under the same expected errors, to offset a slight reduction of PSUs (from 221 to 207) 
# an increase of SSUs involved is observed.
allocation2st_3$expected

#>    Type Dom     V1     V2
#> 1  DOM1   1 0.0900 0.0707
#> 3  DOM1   2 0.0899 0.1101
#> 5  DOM2   1 0.0933 0.0736
#> 7  DOM2   2 0.3370 0.2532
#> 9  DOM2   3 0.3004 0.3240
#> 11 DOM2   4 0.4643 0.4893
#> 13 DOM2   5 0.1769 0.2463
#> 15 DOM2   6 0.1136 0.1362

allocation2st_4$expected

#>    Type Dom     V1     V2
#> 1  DOM1   1 0.0900 0.0659
#> 3  DOM1   2 0.0899 0.1043
#> 5  DOM2   1 0.0929 0.0685
#> 7  DOM2   2 0.3523 0.2406
#> 9  DOM2   3 0.3031 0.3093
#> 11 DOM2   4 0.4780 0.4533
#> 13 DOM2   5 0.1758 0.2354
#> 15 DOM2   6 0.1132 0.1283


## Example 5
# On the contrary, assume to decrease the MINIMUM from 48 to 24.
# The SSUs number strongly decrease in the face of an increase of PSUs,
# always under the same expected errors.
design2 <- design
design2[,4] <- 24
allocation2st_5 <- beat.2st(stratif=strata, errors=errors2,
des_file=design2, psu_file=PSU_strat, rho=rho)

#>   iterations PSU_SR PSU NSR PSU Total   SSU
#> 1          0      0       0         0  8463
#> 2          1     38     161       199 13382
#> 3          2     71     222       293 12471
#> 4          3     69     211       280 12503
#> 5          4     69     212       281 12504

allocation2st_4$expected

#>    Type Dom     V1     V2
#> 1  DOM1   1 0.0900 0.0659
#> 3  DOM1   2 0.0899 0.1043
#> 5  DOM2   1 0.0929 0.0685
#> 7  DOM2   2 0.3523 0.2406
#> 9  DOM2   3 0.3031 0.3093
#> 11 DOM2   4 0.4780 0.4533
#> 13 DOM2   5 0.1758 0.2354
#> 15 DOM2   6 0.1132 0.1283

allocation2st_5$expected

#>    Type Dom     V1     V2
#> 1  DOM1   1 0.0900 0.0842
#> 3  DOM1   2 0.0899 0.1261
#> 5  DOM2   1 0.0934 0.0881
#> 7  DOM2   2 0.3352 0.2862
#> 9  DOM2   3 0.2974 0.3581
#> 11 DOM2   4 0.4478 0.5815
#> 13 DOM2   5 0.1788 0.2765
#> 15 DOM2   6 0.1138 0.1588


# Note that MINIMUM can be different for each stratum.

## Example 6
# Assume that the SSUs are in turn clusters, for instance households composed by individuals.
# In the previous examples we always derived optimal allocations
# for sample of SSUs (i.e. households, because
# DELTA = 1).
design

#>    STRATUM STRAT_MOS DELTA MINIMUM
#> 1        1   2302353     1      48
#> 2        2    179562     1      48
#> 3        3    371802     1      48
#> 4        4    592303     1      48
#> 5        5    221687     1      48
#> 6        6    440613     1      48
#> 7        7    128230     1      48
#> 8        8    222648     1      48
#> 9        9    287906     1      48
#> 10      10    882718     1      48
#> 11      11    223516     1      48
#> 12      12    883285     1      48
#> 13      13    594988     1      48
#> 14      14    183169     1      48
#> 15      15   3156694     1      48
#> 16      16   1098740     1      48
#> 17      17   1256025     1      48

design1

#>    STRATUM STRAT_MOS DELTA MINIMUM
#> 1        1   2302353     1      60
#> 2        2    179562     1      60
#> 3        3    371802     1      60
#> 4        4    592303     1      60
#> 5        5    221687     1      60
#> 6        6    440613     1      60
#> 7        7    128230     1      60
#> 8        8    222648     1      60
#> 9        9    287906     1      60
#> 10      10    882718     1      60
#> 11      11    223516     1      60
#> 12      12    883285     1      60
#> 13      13    594988     1      60
#> 14      14    183169     1      60
#> 15      15   3156694     1      60
#> 16      16   1098740     1      60
#> 17      17   1256025     1      60

design2

#>    STRATUM STRAT_MOS DELTA MINIMUM
#> 1        1   2302353     1      24
#> 2        2    179562     1      24
#> 3        3    371802     1      24
#> 4        4    592303     1      24
#> 5        5    221687     1      24
#> 6        6    440613     1      24
#> 7        7    128230     1      24
#> 8        8    222648     1      24
#> 9        9    287906     1      24
#> 10      10    882718     1      24
#> 11      11    223516     1      24
#> 12      12    883285     1      24
#> 13      13    594988     1      24
#> 14      14    183169     1      24
#> 15      15   3156694     1      24
#> 16      16   1098740     1      24
#> 17      17   1256025     1      24

# For obtaining a sample in terms of the elements composing SSUs
# (i.e., individuals) is just sufficient to 
# modify the DELTA in des_file.
design3 <- design
design3$DELTA <- 2.31
# DELTA_IND=2.31, the average size of household in Italy.
allocation2st_6 <- beat.2st(stratif=strata, errors=errors,
des_file=design3, psu_file=PSU_strat, rho=rho)

#>   iterations PSU_SR PSU NSR PSU Total   SSU
#> 1          0      0       0         0  6858
#> 2          1      3      43        46 41547
#> 3          2     48     150       198 29305
#> 4          3     30     133       163 31725
#> 5          4     33     137       170 31178