Starting values for the Design Effect (\(deft\))

Example data frame containing the starting values for the Design Effect (\(deft\)) .

data(beat.example)

Format

The Design Effect data frame contains a row per each stratum with the following variables:

STRATUM

Identifier of the stratum (numeric).

DEFT1

Starting values for the Design Effect in the stratum of the first variable (numeric).

DEFTj

Starting values for the Design Effect in the stratum of the j-th variable (numeric).

DEFTn

Starting values for the Design Effect in the stratum of the last variable (numeric).

Details

Note: the names of the variables must be the ones indicated above.

This is an optional input. The function beat.2st independently computes and updates the design effect. However, it is possible to set the starting values of design effect for each variable in each stratum. The design effect is the square root of the ratio of the actual sampling variance to the variance expected with the simple random sampling (SRS), on equal sample size. Under SRS the desing effect is equal to 1. Usually, as increasing the stages of selection the design effect increases because it takes into account the "clusterization" of sampling units and the sample size in Self Representative (SR) and Non Self Representative (NSR) strata. In practice, higher is the intraclass correlation, higher will be the design effect and much more sample size for satisfying the precision constraints is needed with respect to SRS.

Examples

# Load example data
data(beat.example)
deft_start

#>    STRATUM DEFT1 DEFT2
#> 1        1   1.1   1.1
#> 2        2   1.1   1.1
#> 3        3   1.1   1.1
#> 4        4   1.1   1.1
#> 5        5   1.1   1.1
#> 6        6   1.1   1.1
#> 7        7   1.1   1.1
#> 8        8   1.1   1.1
#> 9        9   1.1   1.1
#> 10      10   1.1   1.1
#> 11      11   1.1   1.1
#> 12      12   1.1   1.1
#> 13      13   1.1   1.1
#> 14      14   1.1   1.1
#> 15      15   1.1   1.1
#> 16      16   1.1   1.1
#> 17      17   1.1   1.1

str(deft_start)

#> 'data.frame':	17 obs. of  3 variables:
#>  $ STRATUM: int  1 2 3 4 5 6 7 8 9 10 ...
#>  $ DEFT1  : num  1.1 1.1 1.1 1.1 1.1 1.1 1.1 1.1 1.1 1.1 ...
#>  $ DEFT2  : num  1.1 1.1 1.1 1.1 1.1 1.1 1.1 1.1 1.1 1.1 ...