The function prepares the main input for the function *beat.1st* in R2BEAT R-package. It returns a list with the following objects: a dataframe, *file_strata*, with the strata information that is one of the inputs for performing the multivariate and multidomainain optimal allocation implemented in the function *beat.1st*; a vector, *var_label*, listing the target variable(s) which will be used for leading the optimal allocation.

The function prepares the main input for the function *beat.1st* in R2BEAT R-package. It returns a list with the following objects: a dataframe, *file_strata*, with the strata information that is one of the inputs for performing the multivariate and multidomainain optimal allocation implemented in the function *beat.1st*; a vector, *var_label*, listing the target variable(s) which will be used for leading the optimal allocation.

prepareInputToAllocation_beat.1st(
  frame,
  ID,
  stratum,
  domain,
  target,
  sample = NULL,
  weights = NULL
)

Arguments

frame:

A dataframe of population data (i.e. population frame or register) containing necessarily the identifier of the units, strata and domainain variables and optionally the target variable(s).

ID:

Name of the identifier of the units in the sampling frame.

stratum:

Either name of the variable in data which is taken as the stratum, or the name of the variables which have to be concatenated to obtain the stratum. In this latter case, the variables used to build the stratum are retained.

domain:

Name of the variable(s) in data which are the domain(s) of the estimates. Domain(s) are aggregations of strata for which the efficiency of the estimates of the target variable(s) will be kept under control when determining the optimal allocation.

target:

Name of variable(s) in data which will be the variables leading the optimal allocation, namely the efficiency of their estimates at domain(s) level will represent a constraint in the optimal allocation. Categorical variables must be encoded as factor previously.

sample:

A dataframe of sample survey data, containing necessarily the strata and domains variables, the target variable(s) and the sampling weights. In this way, statistical summaries of the target variables will be estimated on the sample. Strata and domains variables must be consistent with those defined for frame dataframe. Default is NULL, meaning that population data are used.

weights:

Name of the variable specifying the sampling weights whether sample survey data are used. Default is NULL, meaning that population data are used.

Value

It returns a list with the following objects: a dataframe, *file_strata*, with the strata information that is one of the inputs for performing the multivariate and multidomain optimal allocation implemented in the function *beat.1st* in R2BEAT R-package; a vector, *var_label*, listing the target variable(s) which will be used for leading the optimal allocation.