Intraclass correlation coefficients for self and non self representative in the strata

Example data frame containing intraclass correlation \(\rho\) in Self Representative (SR) and Non Self Representative (NSR) strata.

data(beat.example)

Format

The intraclass correlation coefficienta (\(\rho\)) data frame contains a row per each stratum with the following variables:

STRATUM

Identifier of the stratum (numeric)

RHO_AR1

intraclass correlation of the elementary units for each primary stage unit of the self representing area belonging to the stratum for the first variable.

RHO_ARj

intraclass correlation of the elementary units for each primary stage unit of the self representing area belonging to the stratum for the j-th variable.

RHO_ARn

intraclass correlation of the elementary units for each primary stage unit of the self representing area belonging to the stratum for the n-th variable.

RHO_NAR1

intraclass correlation of the elementary units for each primary stage unit of the non self representing area belonging to the stratum for the first variable.

RHO_NARj

intraclass correlation of the elementary units for each primary stage unit of the non self representing area belonging to the stratum for the j-th variable.

RHO_NARn

intraclass correlation of the elementary units for each primary stage unit of the non self representing area belonging to the stratum for the n-th variable.

Details

Note: the names of the variables must be the ones indicated above.

Intraclass correlation, \(\rho\), provide a measure of the cluster heterogeneity and they have a direct impact on the design effect (design). It can be indirectly computed from the design effect and the average minimum number of interviews in the Primary Stage Units (PSUs). The ideal situation is when all the clusters in which the population is divided are more heterogeneous possible within them. At the limit, if each cluster were a reduced copy of the population then it would be sufficient to extract one just to have the same information that would be obtained from a complete survey. Then, more similar the units in the cluster are, higher the sample size must be (Cochran, 1977, Chapter 8).

By definition, in SR strata \(\rho\), is equal to 1, because there is just a single PSU in SR strata. In NSR strata usually, \(\rho\) is usual higher than 1, because a double stage of selection is needed.

References

Cochran, W. (1977) Sampling Techniques. John Wiley & Sons, Inc., New York.

Examples

if (FALSE) { # \dontrun{
# Load example data
data(beat.example)
rho
str(rho)
} # }