Computation of coefficient of variation (CV) for a given multivariate multiple allocation

Compute the coefficients of variation considering a given multivariate optimal allocation.

beat.cv(n_file, stratif, errors, des_file, psu_file, rho, epsilon)

Arguments

n_file
Data frame containing the sample size allocated in each stratum, for more details, e.g.,allocation.

stratif
Data frame of survey strata, for more details see, e.g.,strata.

errors
Data frame of expected coefficients of variation (CV) for each domain, for more details see, e.g.,errors.

des_file
Data frame containing information on sampling design variables, for more details see, e.g.,design.

psu_file
Data frame containing information on primary stage units stratification, for more details see, e.g.,PSU_strat.

rho
Data frame of survey strata, for more details see, e.g.,rho.

epsilon
The same as in function beat.1st.

Details

This function enables to derive the expected coefficient of variation (CV) from a given allocation. The function beat.cv returns the estimates expected accuracy in terms of coefficient of variation, for several variables in different domains, given a certain allocation among the different strata.

Value

Object of class list. The list contains a set of data.frame, as many of the cross product between domain and interest variables, containing total estimates, population, variance and expected coefficient of variation for every domain modality. For each domain and each variable is defined

Tot1

Total estimate

N

Measure of size

Varfin

The sample variance of the total estimate

CV

The coefficient of variation of the

Examples


# Load example data
data(beat.example)

## Example 1
# Calculate coefficients of variation, for two variables in two domains,
# given an allocation among the different strata.
allocation

#>    SIZE
#> 1  4276
#> 2   350
#> 3    59
#> 4    96
#> 5    42
#> 6    97
#> 7    26
#> 8    45
#> 9    56
#> 10  199
#> 11   41
#> 12  184
#> 13  127
#> 14   42
#> 15  630
#> 16  273
#> 17  315

cv1<-beat.cv( n_file=allocation, stratif=strata, errors=errors,
des_file=design, psu_file=PSU_strat, rho=rho)

## Example 2
# Take the example 1 in beat.2st.
allocation2st_1 <- beat.2st(stratif=strata, errors=errors,
des_file=design, psu_file=PSU_strat,rho=rho)

#>    iterations PSU_SR PSU NSR PSU Total   SSU
#> 1           0      0       0         0  6858
#> 2           1     16      79        95 17170
#> 3           2     42     161       203 14736
#> 4           3     33     153       186 15273
#> 5           4     36     155       191 15147
#> 6           5     33     156       189 15274
#> 7           6     36     155       191 15146
#> 8           7     33     156       189 15273
#> 9           8     36     155       191 15147
#> 10          9     33     156       189 15274
#> 11         10     36     155       191 15146
#> 12         11     33     156       189 15273
#> 13         12     36     155       191 15147
#> 14         13     33     156       189 15274
#> 15         14     36     155       191 15146
#> 16         15     33     156       189 15273
#> 17         16     36     155       191 15147
#> 18         17     33     156       189 15274
#> 19         18     36     155       191 15146
#> 20         19     33     156       189 15273
#> 21         20     36     155       191 15147

# The allocation obtained is 
allocation2st_1$alloc

#>    STRATUM ALLOC  PROP EQUAL
#> 1        1  8900  2677   891
#> 2        2   939   209   891
#> 3        3   147   432   891
#> 4        4   278   689   891
#> 5        5   108   258   891
#> 6        6   219   512   891
#> 7        7    95   149   891
#> 8        8   157   259   891
#> 9        9   137   335   891
#> 10      10   403  1026   891
#> 11      11    65   260   891
#> 12      12   485  1027   891
#> 13      13   270   692   891
#> 14      14    91   213   891
#> 15      15  1322  3671   891
#> 16      16   612  1278   891
#> 17      17   919  1461   891
#> 18   Total 15147 15147 15147

# with these precision constraints 
errors

#>    DOM  CV1  CV2
#> 1 DOM1 0.10 0.99
#> 2 DOM2 0.99 0.99

# and these expacted coefficient of variaction
allocation2st_1$expected

#>    Type Dom     V1     V2
#> 1  DOM1   1 0.1000 0.0768
#> 3  DOM1   2 0.0999 0.1219
#> 5  DOM2   1 0.1033 0.0800
#> 7  DOM2   2 0.3910 0.2749
#> 9  DOM2   3 0.3387 0.3615
#> 11 DOM2   4 0.5255 0.5340
#> 13 DOM2   5 0.1957 0.2724
#> 15 DOM2   6 0.1257 0.1507


# Now, fit the output of beat.2st to allocation, that is
SIZE <- allocation2st_1$alloc[-18,c(2)]
allocation1 <- data.frame(SIZE)
# If apply beat.cv the same error in allocation2st_1$expected should be obtained.
# In fact
cv2<-beat.cv( n_file=allocation1, stratif=strata, errors=errors,
des_file=design, psu_file=PSU_strat, rho=rho)
cv2

#>   Type Dom        V1         V2
#> 1 DOM1   1 0.1002737 0.07688799
#> 2 DOM1   2 0.1006358 0.12215910
#> 3 DOM2   1 0.1032735 0.07997684
#> 4 DOM2   2 0.4019677 0.27834395
#> 5 DOM2   3 0.3386786 0.36147314
#> 6 DOM2   4 0.5254661 0.53398185
#> 7 DOM2   5 0.2015243 0.27465412
#> 8 DOM2   6 0.1256738 0.15067760


# Please, note that some very slightly differences may occur.