Estimator effect

Example data frame containing estimator effect, (\(effst\)), in each stratum for each variable.

data(beat.example)

Format

The estimator effect data frame contains a row per each stratum with the following variables:

STRATUM

Identifier of the stratum (numeric).

EFFST1

Estimator effect in the stratum of the first variable (numeric).

EFFSTj

Estimator effect in the stratum of the j-th variable (numeric).

EFFSTn

Estimator effect in the stratum of the last variable (numeric).

Details

Note: the names of the variables must be the ones indicated above.

The estimator effect, (\(effst\)), provides a measure of the variance inflaction or reduction due to the use of a different estimator from the HT (Horvitz and Thompson, 1952). It is equal to the ratio between the sampling variance of the estimator planned to be used and the sampling variance of the HT.

Then, when the HT is used, (\(effst\)) is equal to 1. However, always more often, different estimators, such as calibration estimator (Deville and Särndal, 1992) or generalized regression estimator GREG (Fuller, 2002 and references therein), are used. Usually this kind of estimators take into account auxiliary variables that enables to increase the accuracy of the estimates, that is, they reduce their errors (CV). Then, their \(effst\) is usually lower than 1.

Therefore, taking into account the estimator effect when planning the survey can help in saving sample size or at least to more properly evatuate the allocation.

References

Deville, J.C., Särndal, C.E. (1992). Calibration estimators in survey sampling. Journal of the American statistical Association, 87(418): 376-38.

Fuller, W.A.. (2002). Regression estimation for survey samples. Survey Methodology 28(1): 5-23.

Horvitz, D.G., Thompson, D.J. (1952) A generalization of sampling without replacement from a finite universe. Journal of the American statistical Association, 47(260): 663-685.

Examples

# Load example data
data(beat.example)
effst

#>    STRATUM EFFST1 EFFST2
#> 1        1    1.3    1.3
#> 2        2    1.3    1.3
#> 3        3    1.3    1.3
#> 4        4    1.3    1.3
#> 5        5    1.3    1.3
#> 6        6    1.3    1.3
#> 7        7    1.3    1.3
#> 8        8    1.3    1.3
#> 9        9    1.3    1.3
#> 10      10    1.3    1.3
#> 11      11    1.3    1.3
#> 12      12    1.3    1.3
#> 13      13    1.3    1.3
#> 14      14    1.3    1.3
#> 15      15    1.3    1.3
#> 16      16    1.3    1.3
#> 17      17    1.3    1.3

str(effst)

#> 'data.frame':	17 obs. of  3 variables:
#>  $ STRATUM: int  1 2 3 4 5 6 7 8 9 10 ...
#>  $ EFFST1 : num  1.3 1.3 1.3 1.3 1.3 1.3 1.3 1.3 1.3 1.3 ...
#>  $ EFFST2 : num  1.3 1.3 1.3 1.3 1.3 1.3 1.3 1.3 1.3 1.3 ...