Plot of the sensitivity analysis for some parameters by means of grid search

This function allows to plot the results of the simulation carried out by using the function 'sensitivity'.

plot_sens ( 
  x,
  min,
  max)

Arguments

x

The result of the 'sensitivity' function.

min

minimum value of the parameter.

max

maximum value of the parameter.

Value

A list containing the (i) vector of allocated PSUs in the iterations and (ii) the vector of allocated SSUs in the iterations

Author

Giulio Barcaroli

Examples

if (FALSE) { # \dontrun{
library(readr)
pop <- read_rds("https://github.com/barcaroli/R2BEAT_workflows/blob/master/pop.RDS?raw=true")
library(R2BEAT)
cv <- as.data.frame(list(DOM=c("DOM1","DOM2"),
                         CV1=c(0.02,0.03),
                         CV2=c(0.03,0.05),
                         CV3=c(0.03,0.05),
                         CV4=c(0.04,0.08)))
cv
# parameters
frame <- pop
errors <- cv
id_PSU <- "municipality"  
id_SSU <- "id_ind"        
stratum <- "stratum"  
target <- c("income_hh","active","inactive","unemployed")   # more than one
deff_level <- "stratum"        
domain <- "region"        
minimum <- 50 # minimum number of SSUs to be interviewed in each selected PSU
# average dimension of the SSU in terms of elementary survey units
#delta =  nrow(pop) /length(unique(pop$id_hh))     
delta =  1  # average dimension of the SSU in terms of elementary survey units
deff_sugg <- 1.5
min <- 30
max <- 80
sens <- sensitivity_min_SSU (
  frame,
  errors,
  id_PSU,
  id_SSU,
  stratum,
  target,
  deff_level,
  domain,
  minimum,
  delta,
  deff_sugg,
  min,
  max)
plot_sens(sens,min,max)
} # }